services - Services adicionados v3

+++ src/main/java/br/com/projeto/labpcp/service/AlunoService.java
@@ -1,9 +1,12 @@
 package br.com.projeto.labpcp.service;
 
 
+import br.com.projeto.labpcp.controller.dto.request.AlterarAlunoRequest;
+import br.com.projeto.labpcp.controller.dto.request.AlterarDocenteRequest;
 import br.com.projeto.labpcp.controller.dto.request.InserirAlunoRequest;
 import br.com.projeto.labpcp.controller.dto.response.AlunoResponse;
 import br.com.projeto.labpcp.datasource.entity.AlunoEntity;
+import br.com.projeto.labpcp.datasource.entity.DocenteEntity;
 import br.com.projeto.labpcp.datasource.entity.TurmaEntity;
 import br.com.projeto.labpcp.datasource.entity.UsuarioEntity;
 import br.com.projeto.labpcp.datasource.repository.AlunoRepository;
@@ -38,12 +41,14 @@ public class AlunoService {
                 tokenService.buscaCampo(token, "sub")
         );
 
-        UsuarioEntity usuario = usuarioRepository.findById(idUsuario)
+        usuarioRepository.findById(idUsuario)
                 .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
 
 
+        UsuarioEntity usuarioAluno = usuarioRepository.findById(inserirAlunoRequest.id()).orElseThrow(() -> new RuntimeException("Id não encontrado"));
+
         AlunoEntity alunoEntity = new AlunoEntity();
-        alunoEntity.setUsuario(usuario);
+        alunoEntity.setUsuario(usuarioAluno);
         alunoEntity.setNome(inserirAlunoRequest.nome());
         alunoEntity.setDataNascimento(inserirAlunoRequest.dataNascimento());
 
@@ -58,17 +63,55 @@ public class AlunoService {
 
     }
 
+    public AlunoEntity buscarPorId(Long id, String token) {
+
+        String nomePapel = tokenService.buscaCampo(token, "scope");
+        if (!Objects.equals(nomePapel, "PEDAGAGICO")
+                && !Objects.equals(nomePapel, "ADMIN")) {
+            throw new RuntimeException("Usuário não tem acesso a essa funcionalidade");
+        }
+
+
+        return alunoRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("ID não encontrado"));
+    }
+
+
     public AlunoEntity buscarPorId(Long id) {
         return alunoRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("ID não encontrado"));
-
     }
 
-    public List<AlunoEntity> buscarTodos() {
+
+    public List<AlunoEntity> buscarTodos(String token) {
+        String nomePapel = tokenService.buscaCampo(token, "scope");
+
+        if (!Objects.equals(nomePapel, "PEDAGAGICO")
+                && !Objects.equals(nomePapel, "ADMIN")) {
+            throw new RuntimeException("Usuário não tem acesso a essa funcionalidade");
+        }
+
         return alunoRepository.findAll();
     }
 
-    public void excluirPorId(Long id) {
+    public void excluirPorId(Long id, String token) {
+        String nomePapel = tokenService.buscaCampo(token, "scope");
+
+        if (!Objects.equals(nomePapel, "PEDAGAGICO")
+                && !Objects.equals(nomePapel, "ADMIN")) {
+            throw new RuntimeException("Usuário não tem acesso a essa funcionalidade");
+        }
         alunoRepository.deleteById(id);
+    }
+
+
+    public AlunoEntity alterar(Long id, AlterarAlunoRequest alterarAlunoRequest, String token) {
+        String nomePapel = tokenService.buscaCampo(token, "scope");
+        if (!Objects.equals(nomePapel, "PEDAGAGICO")
+                && !Objects.equals(nomePapel, "ADMIN")) {
+            throw new RuntimeException("Usuário não tem acesso a essa funcionalidade");
+        }
+        alunoRepository.updateAlunoParcial(id, alterarAlunoRequest.nome(), alterarAlunoRequest.dataNascimento());
+        return alunoRepository.findById(id).orElseThrow(() -> new RuntimeException("Não foi possível recuperar o usuário"));
     }
 }
